* Removed allAdoptables * Renamed pagination_adoptables to adoptable * database only gets filled with data if on development environment * Added a lot of values that should be in env variables to env file

+++ main.py
@@ -4,6 +4,11 @@ from mongoengine import connect
 from flask_graphql import GraphQLView
 from lib.schema.AdoptableSchema import schema
 from populate_db import populate
+from dotenv import load_dotenv
+from os import getenv
+
+load_dotenv()
+
 
 app = Flask(__name__)
 CORS(app)
@@ -17,7 +22,8 @@ app.add_url_rule('/graphql', view_func=GraphQLView.as_view(
 
 
 if __name__ == '__main__':
-    db = connect('RepoAdopt', host='mongodb://localhost/', port=27017, alias='default')
-    db.drop_database('RepoAdopt')
-    populate()
-    app.run(debug=True, host='0.0.0.0', port=5000)
+    db = connect('RepoAdopt', host=getenv("HOST"), port=getenv("DBPORT"), alias='default')
+    if getenv("ENVIRONMENT") == "develop":
+        db.drop_database('RepoAdopt')
+        populate()
+    app.run(debug=True, host='0.0.0.0', port=getenv("PORT"))
